+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 
+## v1.0.4
+* Changed: Fixed errors for `MinVoltageCellId`, `MinCellVoltage`, `MaxVoltageCellId` and `MaxCellVoltage` calculations
+* Changed: Improved error handling and output
+
 ## v1.0.3
 * Added: Check if provided values are string, integer or float
-* Changed: Time-To-Go can now be configured in the config file. When charging the battery it's calculating the time to 100% SoC and when discharging it's calculating the time to the SoC configured in the config file.
+* Changed: Time-To-Go can now be configured in the config file. When charging the battery it's calculating the time to 100% SoC and when discharging it's calculating the time to the SoC configured in the config file
 
 ## v1.0.2
 * Added: Timeout in order to disconnect the battery, if no new MQTT message is received after x seconds (configurable in `config.ini`)